pyproject.toml
[project]
name = "beammodebot"
version = "0.1.0"
description = "Clone of tg @HowGayBot"
authors = [
    {name = "idkravitz",email = "idkravitz@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "telethon (>=1.42.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "mmh3 (>=5.2.0,<6.0.0)"
]

[project.scripts]
beammodebot = "beammodebot.main:main"

[tool.poetry]
packages = [{include = "beammodebot", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
